backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
sqlalchemy
psycopg2-binary